requirements.txt
Django==4.2.6
djangorestframework==3.15.0
djangorestframework-simplejwt==5.2.2
pymongo==4.5.0
